% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Prints hello to screen}
\usage{
hello(name)
}
\arguments{
\item{name}{A character vector of length 1}
}
\value{
Returns the hello string
}
\description{
This is a simple function that nobody asked for.
}
